Change downloads emit via loop
The previous style of the content_unit_downloader caused the last
content unit to be emitted as None. There was no other way to signal to
the plugin writer that there are no more content units for them.

A loop can automatically use StopIteration() to solve this problem,
which greatly shortens the amount of code a plugin writer has to write.

This also eliminates the need for the plugin writer to instantiate the
asyncio event loop on their own, so that is done for them now.